requirements.txt
httpx>=0.27
websockets>=12.0
polars>=1.0
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.7
pydantic-settings>=2.3
python-json-logger>=2.0
tenacity>=8.3
sqlalchemy>=2.0
aiosqlite>=0.20
asyncpg>=0.29
redis>=5.0
python-telegram-bot>=21.0
PyJWT>=2.8
orjson>=3.10
typer>=0.12
